Improve numeric types.
Don't store fractional meters for banner lengths.
Store coordinates with 6 digits.

+++ src/main/java/com/bannergress/backend/dto/BannerDto.java
@@ -67,7 +67,7 @@ public class BannerDto {
     /**
      * Length in meters.
      */
-    public Double lengthMeters;
+    public Integer lengthMeters;
 
     /**
      * Address.
